@@ -10,13 +10,13 @@
     "@foxglove/schemas": "0.7.0",
     "@foxglove/studio": "workspace:*",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.4",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.13",
+    "@types/react": "18.0.26",
+    "@types/react-dom": "18.0.9",
     "clean-webpack-plugin": "4.0.0",
     "html-webpack-plugin": "5.5.0",
     "mathjs": "10.1.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "webpack": "5.74.0",
     "webpack-dev-server": "4.11.0"
   }
