@@ -42,8 +42,8 @@
     "plist": "3.0.5",
     "pngjs": "6.0.0",
     "quicklookjs": "0.0.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-refresh-typescript": "2.0.3",
     "react-use": "17.4.0",
     "tinycolor2": "1.4.2",
