@@ -20,7 +20,7 @@
     "prepack": "tsc -b"
   },
   "devDependencies": {
-    "@testing-library/react-hooks": "7.0.2",
+    "@testing-library/react": "13.4.0",
     "typescript": "4.8.2"
   },
   "dependencies": {
@@ -28,6 +28,6 @@
     "shallowequal": "1.1.0"
   },
   "peerDependencies": {
-    "react": "^17.0.2"
+    "react": "^18.2.0"
   }
 }
