@@ -68,10 +68,9 @@
     "@sentry/types": "7.10.0",
     "@storybook/addon-actions": "6.5.15",
     "@storybook/react": "6.5.15",
-    "@testing-library/dom": "8.11.3",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/react-hooks": "7.0.2",
-    "@testing-library/user-event": "14.3.0",
+    "@testing-library/dom": "8.19.0",
+    "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "14.4.3",
     "@types/amplitude-js": "8.16.2",
     "@types/base16": "^1",
     "@types/circular-dependency-plugin": "5.0.5",
@@ -86,9 +85,9 @@
     "@types/nearley": "2.11.2",
     "@types/offscreencanvas": "2019.6.4",
     "@types/prettier": "2.4.4",
-    "@types/react": "17.0.39",
+    "@types/react": "18.0.26",
     "@types/react-autocomplete": "1.8.6",
-    "@types/react-dom": "17.0.13",
+    "@types/react-dom": "18.0.9",
     "@types/react-hover-observer": "workspace:*",
     "@types/react-table": "7.7.9",
     "@types/react-transition-group": "^4.4.4",
@@ -152,17 +151,17 @@
     "notistack": "2.0.5",
     "path-browserify": "1.0.1",
     "prettier": "2.5.1",
-    "react": "17.0.2",
+    "react": "18.2.0",
     "react-autocomplete": "https://github.com/foxglove/react-autocomplete/raw/9717d597d7c78c316fb7ad3c2c57060c243530c2/react-autocomplete-9.9.9.tgz",
     "react-colorful": "5.6.0",
     "react-dnd": "14.0.2",
     "react-dnd-html5-backend": "14.0.2",
-    "react-dom": "17.0.2",
+    "react-dom": "18.2.0",
     "react-hover-observer": "2.1.1",
     "react-json-tree": "patch:react-json-tree@npm:0.15.1#../../patches/react-json-tree.patch",
     "react-markdown": "8.0.0",
     "react-monaco-editor": "0.46.0",
-    "react-mosaic-component": "5.0.0",
+    "react-mosaic-component": "5.3.0",
     "react-refresh-typescript": "2.0.3",
     "react-resize-detector": "7.1.2",
     "react-svg-loader": "3.0.3",
