@@ -17,8 +17,8 @@
     "copy-webpack-plugin": "11.0.0",
     "html-webpack-plugin": "5.5.0",
     "idb": "7.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "webpack": "5.74.0",
     "webpack-dev-server": "4.11.0"
   }
