@@ -7,7 +7,7 @@ import LeafletRetinaIconUrl from "leaflet/dist/images/marker-icon-2x.png";
 import LeafletIconUrl from "leaflet/dist/images/marker-icon.png";
 import LeafletShadowIconUrl from "leaflet/dist/images/marker-shadow.png";
 import { StrictMode } from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import { PanelExtensionContext } from "@foxglove/studio";
 
@@ -29,13 +29,13 @@ L.Marker.prototype.options.icon = L.icon({
 });
 
 export function initPanel(context: PanelExtensionContext): () => void {
-  ReactDOM.render(
+  const root = createRoot(context.panelElement);
+  root.render(
     <StrictMode>
       <MapPanel context={context} />
     </StrictMode>,
-    context.panelElement,
   );
   return () => {
-    ReactDOM.unmountComponentAtNode(context.panelElement);
+    root.unmount();
   };
 }
