@@ -81,7 +81,7 @@ function makeUnsavedLayout(id: number): Layout {
 async function clickMenuButtonAction(index: number) {
   const actions = await screen.findAllByTestId("layout-actions");
   if (actions[index]) {
-    fireEvent.click(actions[index]!);
+    await userEvent.click(actions[index]!);
   }
 }
 
@@ -89,16 +89,16 @@ async function deleteLayoutInteraction(index: number) {
   await clickMenuButtonAction(index);
 
   const deleteButton = await screen.findByText("Delete");
-  fireEvent.click(deleteButton);
+  await userEvent.click(deleteButton);
   const confirmButton = await screen.findByText("Delete");
-  fireEvent.click(confirmButton);
+  await userEvent.click(confirmButton);
 }
 
 async function doMultiAction(action: string) {
   await selectAllAction();
   await clickMenuButtonAction(0);
   const button = await screen.findByText(action);
-  fireEvent.click(button);
+  await userEvent.click(button);
 }
 
 async function selectAllAction() {
@@ -196,12 +196,15 @@ MultiSelect.play = async () => {
 export function MultiDelete(): JSX.Element {
   return <LayoutBrowser />;
 }
-MultiDelete.parameters = { colorScheme: "dark" };
+MultiDelete.parameters = {
+  colorScheme: "dark",
+  chromatic: { disableSnapshot: true }, // FG-1083
+};
 MultiDelete.play = async () => {
   await doMultiAction("Delete");
 
   const confirmButton = await screen.findByText("Delete");
-  fireEvent.click(confirmButton);
+  await userEvent.click(confirmButton);
 };
 
 export function MultiDuplicate(): JSX.Element {
@@ -210,6 +213,7 @@ export function MultiDuplicate(): JSX.Element {
 MultiDuplicate.parameters = {
   colorScheme: "dark",
   mockLayouts: [exampleCurrentLayout, makeUnsavedLayout(1), shortLayout],
+  chromatic: { disableSnapshot: true }, // FG-1083
 };
 MultiDuplicate.play = async () => {
   await doMultiAction("Duplicate");
@@ -221,12 +225,12 @@ export function MultiRevert(): JSX.Element {
 MultiRevert.parameters = {
   colorScheme: "dark",
   mockLayouts: [makeUnsavedLayout(1), makeUnsavedLayout(2), makeUnsavedLayout(3)],
+  chromatic: { disableSnapshot: true }, // FG-1083
 };
 MultiRevert.play = async () => {
   await doMultiAction("Revert");
-
   const revertButton = await screen.findByText("Discard changes");
-  fireEvent.click(revertButton);
+  await userEvent.click(revertButton);
 };
 
 export function MultiSave(): JSX.Element {
@@ -276,7 +280,7 @@ export function AddLayout(_args: unknown): JSX.Element {
 AddLayout.parameters = { colorScheme: "dark" };
 AddLayout.play = async () => {
   const button = await screen.findByTestId("add-layout");
-  fireEvent.click(button);
+  await userEvent.click(button);
 };
 
 export function MenuOpen(_args: unknown): JSX.Element {
@@ -286,7 +290,7 @@ MenuOpen.parameters = { colorScheme: "dark" };
 MenuOpen.play = async () => {
   const actions = await screen.findAllByTestId("layout-actions");
   if (actions[1]) {
-    fireEvent.click(actions[1]);
+    await userEvent.click(actions[1]);
   }
 };
 
@@ -295,7 +299,7 @@ MenuOpenLight.parameters = { colorScheme: "light" };
 MenuOpenLight.play = async () => {
   const actions = await screen.findAllByTestId("layout-actions");
   if (actions[1]) {
-    fireEvent.click(actions[1]);
+    await userEvent.click(actions[1]);
   }
 };
 
@@ -306,10 +310,10 @@ EditingName.parameters = { colorScheme: "dark" };
 EditingName.play = async () => {
   const actions = await screen.findAllByTestId("layout-actions");
   if (actions[1]) {
-    fireEvent.click(actions[1]);
+    await userEvent.click(actions[1]);
   }
   const button = await screen.findByText("Rename");
-  fireEvent.click(button);
+  await userEvent.click(button);
 };
 
 export function CancelRenameWithEscape(_args: unknown): JSX.Element {
@@ -319,10 +323,10 @@ CancelRenameWithEscape.parameters = { colorScheme: "dark" };
 CancelRenameWithEscape.play = async () => {
   const actions = await screen.findAllByTestId("layout-actions");
   if (actions[1]) {
-    fireEvent.click(actions[1]);
+    await userEvent.click(actions[1]);
   }
   const button = await screen.findByText("Rename");
-  fireEvent.click(button);
+  await userEvent.click(button);
   fireEvent.keyDown(document.activeElement!, { key: "Escape" });
 };
 
@@ -333,12 +337,12 @@ CommitRenameWithTab.parameters = { colorScheme: "dark" };
 CommitRenameWithTab.play = async () => {
   const actions = await screen.findAllByTestId("layout-actions");
   if (actions[1]) {
-    fireEvent.click(actions[1]);
+    await userEvent.click(actions[1]);
   }
   const button = await screen.findByText("Rename");
-  fireEvent.click(button);
-  fireEvent.change(document.activeElement!, { target: { value: "New name" } });
-  fireEvent.focusOut(document.activeElement!);
+  await userEvent.click(button);
+  await userEvent.keyboard("New name");
+  await userEvent.tab();
 };
 
 export function Duplicate(_args: unknown): JSX.Element {
@@ -348,16 +352,19 @@ Duplicate.parameters = { colorScheme: "dark" };
 Duplicate.play = async () => {
   const actions = await screen.findAllByTestId("layout-actions");
   if (actions[1]) {
-    fireEvent.click(actions[1]);
+    await userEvent.click(actions[1]);
   }
   const button = await screen.findByText("Duplicate");
-  fireEvent.click(button);
+  await userEvent.click(button);
 };
 
 export function DeleteLayout(_args: unknown): JSX.Element {
   return <LayoutBrowser />;
 }
-DeleteLayout.parameters = { colorScheme: "dark" };
+DeleteLayout.parameters = {
+  colorScheme: "dark",
+  chromatic: { disableSnapshot: true }, // FG-1083
+};
 DeleteLayout.play = async () => await deleteLayoutInteraction(0);
 
 export function DeleteSelectedLayout(_args: unknown): JSX.Element {
@@ -366,14 +373,17 @@ export function DeleteSelectedLayout(_args: unknown): JSX.Element {
 DeleteSelectedLayout.play = async () => {
   const layouts = await screen.findAllByTestId("layout-list-item");
   if (layouts[1]) {
-    fireEvent.click(layouts[1]);
+    await userEvent.click(layouts[1]);
   }
   await deleteLayoutInteraction(1);
   if (layouts[0]) {
-    fireEvent.click(layouts[0]);
+    await userEvent.click(layouts[0]);
   }
 };
-DeleteSelectedLayout.parameters = { colorScheme: "dark" };
+DeleteSelectedLayout.parameters = {
+  colorScheme: "dark",
+  chromatic: { disableSnapshot: true }, // FG-1083
+};
 
 export function DeleteLastLayout(_args: unknown): JSX.Element {
   return <LayoutBrowser />;
@@ -381,5 +391,6 @@ export function DeleteLastLayout(_args: unknown): JSX.Element {
 DeleteLastLayout.parameters = {
   mockLayouts: [exampleCurrentLayout],
   colorScheme: "dark",
+  chromatic: { disableSnapshot: true }, // FG-1083
 };
 DeleteLastLayout.play = async () => await deleteLayoutInteraction(0);
