@@ -14,7 +14,7 @@
 
 /* eslint-disable jest/no-conditional-expect */
 
-import { renderHook, act } from "@testing-library/react-hooks";
+import { renderHook, act } from "@testing-library/react";
 import { PropsWithChildren, useCallback, useState } from "react";
 import { DeepPartial } from "ts-essentials";
 
@@ -38,10 +38,11 @@ import { MAX_PROMISE_TIMEOUT_TIME_MS } from "./pauseFrameForPromise";
 
 jest.setTimeout(MAX_PROMISE_TIMEOUT_TIME_MS * 3);
 
-// We require two state updates for each player emit() to take effect, because we  React 18 / @testing-library/react,
+// We require two act() calls for each player emit() to take effect because of some quirks with
+// React 18 / @testing-library/react. The root cause for why this requires two act() calls is unknown.
 async function doubleAct(fn: () => Promise<void>) {
   let promise: Promise<void> | undefined;
-  act(() => void (promise = fn()));
+  await act(() => void (promise = fn()));
   await act(async () => await promise);
 }
 
@@ -272,7 +273,7 @@ describe("MessagePipelineProvider/useMessagePipeline", () => {
     renderHook(Hook, {
       wrapper: Wrapper,
     });
-    act(() => {
+    await act(() => {
       void player.emit();
     });
     await expect(async () => await player.emit()).rejects.toThrow(
@@ -287,12 +288,12 @@ describe("MessagePipelineProvider/useMessagePipeline", () => {
       wrapper: Wrapper,
     });
 
-    act(() => {
+    await act(() => {
       result.current.setSubscriptions("test", [{ topic: "/studio/test" }]);
     });
     expect(result.current.subscriptions).toEqual([{ topic: "/studio/test" }]);
 
-    act(() => {
+    await act(() => {
       result.current.setSubscriptions("bar", [{ topic: "/studio/test2" }]);
     });
     expect(result.current.subscriptions).toEqual([
@@ -342,7 +343,7 @@ describe("MessagePipelineProvider/useMessagePipeline", () => {
         }),
     );
 
-    act(() => {
+    await act(() => {
       result.current.setSubscriptions("custom-id", [{ topic: "/input/foo" }]);
     });
     expect(result.current.subscriptions).toEqual([{ topic: "/input/foo" }]);
@@ -398,7 +399,7 @@ describe("MessagePipelineProvider/useMessagePipeline", () => {
         }),
     );
 
-    act(() => {
+    await act(() => {
       result.current.setSubscriptions("custom-id", [{ topic: "/input/foo" }]);
     });
     expect(result.current.subscriptions).toEqual([{ topic: "/input/foo" }]);
@@ -426,15 +427,14 @@ describe("MessagePipelineProvider/useMessagePipeline", () => {
     rerender();
 
     // Unsubscribe and re-subscribe to trigger injection of old messages
-    act(() => {
+    await act(() => {
       result.current.setSubscriptions("custom-id", []);
     });
-    act(() => {
+    await act(() => {
       result.current.setSubscriptions("custom-id", [{ topic: "/input/foo" }]);
     });
     expect(result.current.subscriptions).toEqual([{ topic: "/input/foo" }]);
-
-    await act(
+    await doubleAct(
       async () =>
         await player2.emit({
           activeData: {
@@ -464,14 +464,14 @@ describe("MessagePipelineProvider/useMessagePipeline", () => {
       wrapper: Wrapper,
     });
 
-    act(() =>
+    await act(() =>
       result.current.setPublishers("test", [{ topic: "/studio/test", schemaName: "test" }]),
     );
     expect(result.current.publishers).toEqual<typeof result.current.publishers>([
       { topic: "/studio/test", schemaName: "test" },
     ]);
 
-    act(() =>
+    await act(() =>
       result.current.setPublishers("bar", [{ topic: "/studio/test2", schemaName: "test2" }]),
     );
     expect(result.current.publishers).toEqual<typeof result.current.publishers>([
@@ -647,9 +647,9 @@ describe("MessagePipelineProvider/useMessagePipeline", () => {
     const { result, rerender } = renderHook(Hook, {
       wrapper: Wrapper,
     });
-    act(() => result.current.setSubscriptions("test", [{ topic: "/studio/test" }]));
-    act(() => result.current.setSubscriptions("bar", [{ topic: "/studio/test2" }]));
-    act(() =>
+    await act(() => result.current.setSubscriptions("test", [{ topic: "/studio/test" }]));
+    await act(() => result.current.setSubscriptions("bar", [{ topic: "/studio/test2" }]));
+    await act(() =>
       result.current.setPublishers("test", [{ topic: "/studio/test", schemaName: "test" }]),
     );
 
@@ -822,7 +822,7 @@ describe("MessagePipelineProvider/useMessagePipeline", () => {
       const firstPlayerResumeFn = result.current.pauseFrame("");
 
       // Then trigger the next emit.
-      act(() => void player.emit());
+      await act(() => void player.emit());
       await delay(20);
 
       // Replace the player.
