@@ -12,8 +12,7 @@
 //   found at http://www.apache.org/licenses/LICENSE-2.0
 //   You may not use this file except in compliance with the License.
 
-import { render } from "@testing-library/react";
-import { renderHook, act } from "@testing-library/react-hooks";
+import { render, act, renderHook } from "@testing-library/react";
 import { useEffect } from "react";
 
 import Panel from "@foxglove/studio-base/components/Panel";
@@ -52,6 +51,7 @@ describe("Panel", () => {
     expect(renderFn.mock.calls).toEqual([
       [{ config: { someString: "hello world" }, saveConfig: expect.any(Function) }],
       [{ config: { someString: "hello world" }, saveConfig: expect.any(Function) }],
+      [{ config: { someString: "hello world" }, saveConfig: expect.any(Function) }],
     ]);
 
     expect(actions).toEqual([
@@ -82,6 +82,7 @@ describe("Panel", () => {
 
     expect(renderFn.mock.calls).toEqual([
       [{ config: { someString }, saveConfig: expect.any(Function) }],
+      [{ config: { someString }, saveConfig: expect.any(Function) }],
     ]);
 
     expect(actions).toEqual([
@@ -109,6 +110,7 @@ describe("Panel", () => {
     );
 
     expect(renderFn.mock.calls).toEqual([
+      [{ config: { someNumber: 42, someString: "hello world" }, saveConfig: expect.any(Function) }],
       [{ config: { someNumber: 42, someString: "hello world" }, saveConfig: expect.any(Function) }],
       [{ config: { someNumber: 42, someString: "hello world" }, saveConfig: expect.any(Function) }],
     ]);
@@ -146,6 +148,7 @@ describe("Panel", () => {
 
     expect(renderFn.mock.calls).toEqual([
       [{ config: { someNumber: 42, someString }, saveConfig: expect.any(Function) }],
+      [{ config: { someNumber: 42, someString }, saveConfig: expect.any(Function) }],
     ]);
 
     expect(actions).toEqual([
@@ -158,7 +161,7 @@ describe("Panel", () => {
     ]);
   });
 
-  it("does not rerender when another panel changes", () => {
+  it("does not rerender when another panel changes", async () => {
     const renderFn = jest.fn();
     const DummyPanel = getDummyPanel(renderFn);
     const childId = "Dummy!1my2ydk";
@@ -174,8 +177,10 @@ describe("Panel", () => {
       },
     });
 
-    expect(renderFn.mock.calls.length).toEqual(2);
-    act(() => actions.current.savePanelConfigs({ configs: [{ id: "someOtherId", config: {} }] }));
-    expect(renderFn.mock.calls.length).toEqual(2);
+    expect(renderFn.mock.calls.length).toEqual(3);
+    await act(() =>
+      actions.current.savePanelConfigs({ configs: [{ id: "someOtherId", config: {} }] }),
+    );
+    expect(renderFn.mock.calls.length).toEqual(3);
   });
 });
