@@ -2,7 +2,7 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
-import { act, renderHook } from "@testing-library/react-hooks";
+import { act, renderHook } from "@testing-library/react";
 import { useLayoutEffect, useRef } from "react";
 
 import MockCurrentLayoutProvider from "@foxglove/studio-base/providers/CurrentLayoutProvider/MockCurrentLayoutProvider";
@@ -10,7 +10,7 @@ import MockCurrentLayoutProvider from "@foxglove/studio-base/providers/CurrentLa
 import { LayoutState, useCurrentLayoutActions, useCurrentLayoutSelector } from "./index";
 
 describe("useCurrentLayoutSelector", () => {
-  it("updates when layout changes", () => {
+  it("updates when layout changes", async () => {
     const { result } = renderHook(
       ({ selector }) => ({
         actions: useCurrentLayoutActions(),
@@ -33,7 +33,7 @@ describe("useCurrentLayoutSelector", () => {
 
     expect(result.current.value).toEqual({ value: 42 });
 
-    act(() => {
+    await act(() => {
       result.current.actions.savePanelConfigs({ configs: [{ id: "foo", config: { value: 1 } }] });
     });
     expect(result.current.value).toEqual({ value: 1 });
