@@ -3,7 +3,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { act, renderHook } from "@testing-library/react-hooks";
+import { act, renderHook } from "@testing-library/react";
 
 import useIndexedDbRecents from "@foxglove/studio-base/hooks/useIndexedDbRecents";
 
@@ -20,7 +20,7 @@ describe("useIndexedDbRecents", () => {
       const { result, unmount } = renderHook(() => useIndexedDbRecents());
       expect(result.current.recents).toEqual([]);
 
-      act(() => {
+      await act(() => {
         result.current.addRecent({
           sourceId: "foo",
           title: "my-title",
