@@ -176,10 +176,11 @@ storiesOf("panels/Table", module)
         <PanelSetup
           fixture={fixture}
           onMount={() => {
-            setImmediate(() => {
+            setImmediate(async () => {
               (
                 document.querySelectorAll("[data-testid=expand-row-0]")[0] as HTMLTableCellElement
               ).click();
+              await Promise.resolve();
               (
                 document.querySelectorAll("[data-testid=expand-row-1]")[0] as HTMLTableCellElement
               ).click();
