@@ -3,7 +3,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { renderHook } from "@testing-library/react-hooks";
+import { renderHook } from "@testing-library/react";
 
 import ConsoleApiContext from "@foxglove/studio-base/context/ConsoleApiContext";
 import CurrentUserContext, { User } from "@foxglove/studio-base/context/CurrentUserContext";
