@@ -3,8 +3,9 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { renderHook } from "@testing-library/react-hooks";
+import { renderHook, act } from "@testing-library/react";
 
+import { signal } from "@foxglove/den/async";
 import { useExtensionCatalog } from "@foxglove/studio-base/context/ExtensionCatalogContext";
 import { ExtensionLoader } from "@foxglove/studio-base/services/ExtensionLoader";
 import { ExtensionInfo } from "@foxglove/studio-base/types/Extensions";
@@ -34,7 +35,11 @@ describe("ExtensionCatalogProvider", () => {
         module.exports = { activate: function() { return 1; } }
     `;
 
-    const loadExtension = jest.fn().mockResolvedValue(source);
+    const loadExtensionCalled = signal();
+    const loadExtension = jest.fn().mockImplementation(async () => {
+      loadExtensionCalled.resolve();
+      return source;
+    });
     const mockPrivateLoader: ExtensionLoader = {
       namespace: "org",
       getExtensions: jest
@@ -45,7 +50,7 @@ describe("ExtensionCatalogProvider", () => {
       uninstallExtension: jest.fn(),
     };
 
-    const { result, waitFor } = renderHook(() => useExtensionCatalog((state) => state), {
+    const { result } = renderHook(() => useExtensionCatalog((state) => state), {
       initialProps: {},
       wrapper: ({ children }) => (
         <ExtensionCatalogProvider loaders={[mockPrivateLoader]}>
@@ -54,7 +59,7 @@ describe("ExtensionCatalogProvider", () => {
       ),
     });
 
-    await waitFor(() => expect(loadExtension).toHaveBeenCalledTimes(1));
+    await act(async () => await loadExtensionCalled);
     expect(result.current.installedExtensions).toEqual([
       {
         description: "description",
@@ -85,7 +90,11 @@ describe("ExtensionCatalogProvider", () => {
         }
     `;
 
-    const loadExtension = jest.fn().mockResolvedValue(source);
+    const loadExtensionCalled = signal();
+    const loadExtension = jest.fn().mockImplementation(async () => {
+      loadExtensionCalled.resolve();
+      return source;
+    });
     const mockPrivateLoader: ExtensionLoader = {
       namespace: "org",
       getExtensions: jest
@@ -96,7 +105,7 @@ describe("ExtensionCatalogProvider", () => {
       uninstallExtension: jest.fn(),
     };
 
-    const { result, waitFor } = renderHook(() => useExtensionCatalog((state) => state), {
+    const { result } = renderHook(() => useExtensionCatalog((state) => state), {
       initialProps: {},
       wrapper: ({ children }) => (
         <ExtensionCatalogProvider loaders={[mockPrivateLoader]}>
@@ -105,7 +114,8 @@ describe("ExtensionCatalogProvider", () => {
       ),
     });
 
-    await waitFor(() => expect(loadExtension).toHaveBeenCalledTimes(1));
+    await act(async () => await loadExtensionCalled);
+    await Promise.resolve();
     expect(result.current.installedMessageConverters).toEqual([
       {
         fromSchemaName: "from.Schema",
