@@ -16,7 +16,9 @@ const muiCache = createCache({ key: "mui", prepend: true });
 // By default the ThemeProvider adds an extra div to the DOM tree. We can disable this with a
 // custom `as` component to FluentThemeProvider. The component must support a `ref` property
 // otherwise we get react warnings.
-const ThemeContainer = React.forwardRef((props, _ref) => <>{props.children}</>);
+const ThemeContainer = React.forwardRef((props: React.PropsWithChildren<unknown>, _ref) => (
+  <>{props.children}</>
+));
 ThemeContainer.displayName = "ThemeContainer";
 
 export default function ThemeProvider({
